+++ NEWS.md
@@ -1,5 +1,6 @@
 ## rtables 0.6.1.9014
  * fix major regressions for page_by machinery caused by migration to formatters 0.5.0 pagination framework.
+ * Allow `page_by = TRUE` in a split with empty values with informative warning.
 
 ## rtables 0.6.1
  * Improved resilience of pagination machinery (`paginate_table`) by generalizing parameters' defaults (`cpp`, `lpp`, and `font_size`).